backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.1
Flask-Migrate==4.1.0
python-dotenv==1.1.0
google-auth==2.40.1
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.166.0
requests==2.32.3
PyJWT==2.10.1
gunicorn==23.0.0